requirements.txt
typer
rich
cryptography
pyyaml